backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.24.0
python-multipart==0.0.6

pydantic==2.9.2
pydantic-settings==2.5.2

bcrypt==4.1.2
PyJWT==2.8.0

pytesseract==0.3.10
Pillow==11.0.0

opencv-python==4.10.0.84

scikit-learn==1.5.2
pandas==2.2.3
numpy==2.1.3

requests==2.31.0
SpeechRecognition==3.10.1
python-dateutil==2.8.2
rapidfuzz==3.10.1
